+++ capa/ida/plugin/view.py
@@ -82,7 +82,9 @@ def parse_node_for_feature(feature, description, comment, depth):
     if feature.startswith("#"):
         display += f"{' ' * depth}{feature}\n"
     elif description:
-        if feature.startswith(("- and", "- or", "- optional", "- basic block", "- not", "- instruction:")):
+        if feature.startswith(
+            ("- and", "- or", "- optional", "- basic block", "- not", "- instruction:")
+        ):
             display += f"{' ' * depth}{feature}\n"
             if comment:
                 display += f" # {comment}"
@@ -254,7 +256,10 @@ class CapaExplorerRulegenPreview(QtWidgets.QTextEdit):
                     # user Tab, indent selected lines
                     lines_modified = end_lineno - start_lineno
                     first_modified = True
-                    change = [self.INDENT + line for line in plain[start_lineno : end_lineno + 1]]
+                    change = [
+                        self.INDENT + line
+                        for line in plain[start_lineno : end_lineno + 1]
+                    ]
                 else:
                     # user SHIFT + Tab, dedent selected lines
                     lines_modified = 0
@@ -279,7 +284,9 @@ class CapaExplorerRulegenPreview(QtWidgets.QTextEdit):
                 if e.key() == QtCore.Qt.Key_Tab:
                     # user Tab, increase increment selection positions
                     select_start_ppos += len(self.INDENT)
-                    select_end_ppos += (lines_modified * len(self.INDENT)) + len(self.INDENT)
+                    select_end_ppos += (lines_modified * len(self.INDENT)) + len(
+                        self.INDENT
+                    )
                 elif lines_modified:
                     # user SHIFT + Tab, decrease selection positions
                     if start_lineco not in (0, 1) and first_modified:
@@ -288,7 +295,9 @@ class CapaExplorerRulegenPreview(QtWidgets.QTextEdit):
                     select_end_ppos -= lines_modified * len(self.INDENT)
 
                 # apply updated selection and restore previous scroll position
-                self.set_selection(select_start_ppos, select_end_ppos, len(self.toPlainText()))
+                self.set_selection(
+                    select_start_ppos, select_end_ppos, len(self.toPlainText())
+                )
                 self.verticalScrollBar().setSliderPosition(scroll_ppos)
         else:
             super().keyPressEvent(e)
@@ -326,7 +335,9 @@ class CapaExplorerRulegenEditor(QtWidgets.QTreeWidget):
         self.setEditTriggers(QtWidgets.QAbstractItemView.NoEditTriggers)
         self.setContextMenuPolicy(QtCore.Qt.CustomContextMenu)
         self.setSelectionMode(QtWidgets.QAbstractItemView.ExtendedSelection)
-        self.setStyleSheet("QTreeView::item {padding-right: 15 px;padding-bottom: 2 px;}")
+        self.setStyleSheet(
+            "QTreeView::item {padding-right: 15 px;padding-bottom: 2 px;}"
+        )
 
         # configure view columns to auto-resize
         for idx in range(3):
@@ -382,9 +393,9 @@ class CapaExplorerRulegenEditor(QtWidgets.QTreeWidget):
         """ """
         super().dragMoveEvent(e)
 
-    def dragEventEnter(self, e):
+    def dragEnterEvent(self, e):
         """ """
-        super().dragEventEnter(e)
+        super().dragEnterEvent(e)
 
     def dropEvent(self, e):
         """ """
@@ -424,7 +435,9 @@ class CapaExplorerRulegenEditor(QtWidgets.QTreeWidget):
         # we don't want to add new features under the invisible root because capa rules should
         # contain a single top-level node; this may not always be the case so we default to the last
         # child node that was added to the invisible root
-        top_node = self.invisibleRootItem().child(self.invisibleRootItem().childCount() - 1)
+        top_node = self.invisibleRootItem().child(
+            self.invisibleRootItem().childCount() - 1
+        )
 
         # create a new parent under top-level node
         new_parent = self.new_expression_node(top_node, (action.data()[0], ""))
@@ -478,7 +491,10 @@ class CapaExplorerRulegenEditor(QtWidgets.QTreeWidget):
         if not self.indexAt(pos).isValid():
             # user selected invalid index
             self.load_custom_context_menu_invalid_index(pos)
-        elif self.itemAt(pos).capa_type == CapaExplorerRulegenEditor.get_node_type_expression():
+        elif (
+            self.itemAt(pos).capa_type
+            == CapaExplorerRulegenEditor.get_node_type_expression()
+        ):
             # user selected expression node
             self.load_custom_context_menu_expression(pos)
         else:
@@ -510,8 +526,12 @@ class CapaExplorerRulegenEditor(QtWidgets.QTreeWidget):
             rule_text += "\n  features:\n"
 
         for o in iterate_tree(self):
-            feature, description, comment = (o.strip() for o in tuple(o.text(i) for i in range(3)))
-            rule_text += parse_node_for_feature(feature, description, comment, calc_item_depth(o))
+            feature, description, comment = (
+                o.strip() for o in tuple(o.text(i) for i in range(3))
+            )
+            rule_text += parse_node_for_feature(
+                feature, description, comment, calc_item_depth(o)
+            )
 
         # TODO(mike-hunhoff): we avoid circular update by disabling signals when updating
         # the preview. Preferably we would refactor the code to avoid this
@@ -546,7 +566,9 @@ class CapaExplorerRulegenEditor(QtWidgets.QTreeWidget):
 
         # build submenu with modify actions
         sub_menu = build_context_menu(self.parent(), sub_actions)
-        sub_menu.setTitle(f"Nest feature{'' if len(tuple(self.get_features(selected=True))) == 1 else 's'}")
+        sub_menu.setTitle(
+            f"Nest feature{'' if len(tuple(self.get_features(selected=True))) == 1 else 's'}"
+        )
 
         # build main menu with submenu + main actions
         menu = build_context_menu(self.parent(), (sub_menu,) + actions)
@@ -562,8 +584,16 @@ class CapaExplorerRulegenEditor(QtWidgets.QTreeWidget):
             ("or", ("- or:", self.itemAt(pos)), self.slot_edit_expression),
             ("not", ("- not:", self.itemAt(pos)), self.slot_edit_expression),
             ("optional", ("- optional:", self.itemAt(pos)), self.slot_edit_expression),
-            ("basic block", ("- basic block:", self.itemAt(pos)), self.slot_edit_expression),
-            ("instruction", ("- instruction:", self.itemAt(pos)), self.slot_edit_expression),
+            (
+                "basic block",
+                ("- basic block:", self.itemAt(pos)),
+                self.slot_edit_expression,
+            ),
+            (
+                "instruction",
+                ("- instruction:", self.itemAt(pos)),
+                self.slot_edit_expression,
+            ),
         )
 
         # build submenu with modify actions
@@ -653,7 +683,9 @@ class CapaExplorerRulegenEditor(QtWidgets.QTreeWidget):
         # we don't want to add new features under the invisible root because capa rules should
         # contain a single top-level node; this may not always be the case so we default to the last
         # child node that was added to the invisible root
-        top_node = self.invisibleRootItem().child(self.invisibleRootItem().childCount() - 1)
+        top_node = self.invisibleRootItem().child(
+            self.invisibleRootItem().childCount() - 1
+        )
 
         # build feature counts
         counted = list(zip(Counter(features).keys(), Counter(features).values()))
@@ -719,15 +751,32 @@ class CapaExplorerRulegenEditor(QtWidgets.QTreeWidget):
             node.setText(idx, text)
 
         # we need to set our own type so we can control the GUI accordingly
-        if feature.startswith(("- and:", "- or:", "- not:", "- basic block:", "- instruction:", "- optional:")):
-            setattr(node, "capa_type", CapaExplorerRulegenEditor.get_node_type_expression())
+        if feature.startswith(
+            (
+                "- and:",
+                "- or:",
+                "- not:",
+                "- basic block:",
+                "- instruction:",
+                "- optional:",
+            )
+        ):
+            setattr(
+                node, "capa_type", CapaExplorerRulegenEditor.get_node_type_expression()
+            )
         elif feature.startswith("#"):
-            setattr(node, "capa_type", CapaExplorerRulegenEditor.get_node_type_comment())
+            setattr(
+                node, "capa_type", CapaExplorerRulegenEditor.get_node_type_comment()
+            )
         else:
-            setattr(node, "capa_type", CapaExplorerRulegenEditor.get_node_type_feature())
+            setattr(
+                node, "capa_type", CapaExplorerRulegenEditor.get_node_type_feature()
+            )
 
         # format the node based on its type
-        (self.set_expression_node, self.set_feature_node, self.set_comment_node)[node.capa_type](node)
+        (self.set_expression_node, self.set_feature_node, self.set_comment_node)[
+            node.capa_type
+        ](node)
 
         parent.addChild(node)
 
@@ -741,7 +790,9 @@ class CapaExplorerRulegenEditor(QtWidgets.QTreeWidget):
         if -1 == rule_text.find("features:"):
             return
 
-        rule_features = rule_text[rule_text.find("features:") + len("features:") :].strip("\n")
+        rule_features = rule_text[
+            rule_text.find("features:") + len("features:") :
+        ].strip("\n")
 
         if not rule_features:
             # no features; nothing to do
@@ -769,7 +820,9 @@ class CapaExplorerRulegenEditor(QtWidgets.QTreeWidget):
                     parent = o
                     break
 
-            node = self.make_child_node_from_feature(parent, parse_yaml_line(line.strip()))
+            node = self.make_child_node_from_feature(
+                parent, parse_yaml_line(line.strip())
+            )
 
             # append our new node in case it's a parent for another node
             if node:
@@ -803,7 +856,10 @@ class CapaExplorerRulegenEditor(QtWidgets.QTreeWidget):
     def get_expressions(self, selected=False, ignore=()):
         """ """
         for expression in filter(
-            lambda o: o.capa_type == CapaExplorerRulegenEditor.get_node_type_expression(), tuple(iterate_tree(self))
+            lambda o: (
+                o.capa_type == CapaExplorerRulegenEditor.get_node_type_expression()
+            ),
+            tuple(iterate_tree(self)),
         ):
             if expression in ignore:
                 continue
@@ -821,7 +877,9 @@ class CapaExplorerRulegenFeatures(QtWidgets.QTreeWidget):
         self.editor = editor
 
         self.setHeaderLabels(["Feature", "Address"])
-        self.setStyleSheet("QTreeView::item {padding-right: 15 px;padding-bottom: 2 px;}")
+        self.setStyleSheet(
+            "QTreeView::item {padding-right: 15 px;padding-bottom: 2 px;}"
+        )
 
         # configure view columns to auto-resize
         for idx in range(2):
@@ -875,7 +933,10 @@ class CapaExplorerRulegenFeatures(QtWidgets.QTreeWidget):
 
     def slot_add_n_bytes_feature(self, action):
         """ """
-        count = idaapi.ask_long(16, f"Enter number of bytes (1-{capa.features.common.MAX_BYTES_FEATURE_SIZE}):")
+        count = idaapi.ask_long(
+            16,
+            f"Enter number of bytes (1-{capa.features.common.MAX_BYTES_FEATURE_SIZE}):",
+        )
         if count and 1 <= count <= capa.features.common.MAX_BYTES_FEATURE_SIZE:
             item = self.selectedItems()[0].data(0, 0x100)
             item.value = item.value[:count]
@@ -892,7 +953,9 @@ class CapaExplorerRulegenFeatures(QtWidgets.QTreeWidget):
 
         if selected_items_count == 1:
             action_add_features_fmt = "Add feature"
-            if isinstance(self.selectedItems()[0].data(0, 0x100), capa.features.common.Bytes):
+            if isinstance(
+                self.selectedItems()[0].data(0, 0x100), capa.features.common.Bytes
+            ):
                 actions.append(("Add n bytes...", (), self.slot_add_n_bytes_feature))
         else:
             action_add_features_fmt = f"Add {selected_items_count} features"
@@ -904,7 +967,9 @@ class CapaExplorerRulegenFeatures(QtWidgets.QTreeWidget):
 
     def slot_item_double_clicked(self, o, column):
         """ """
-        if column == CapaExplorerRulegenFeatures.get_column_address_index() and o.text(column):
+        if column == CapaExplorerRulegenFeatures.get_column_address_index() and o.text(
+            column
+        ):
             idc.jumpto(int(o.text(column), 0x10))
         elif o.capa_type == CapaExplorerRulegenFeatures.get_node_type_leaf():
             self.editor.update_features([o.data(0, 0x100)])
@@ -1033,9 +1098,14 @@ class CapaExplorerRulegenFeatures(QtWidgets.QTreeWidget):
 
     def load_features(self, file_features, func_features: Optional[dict] = None):
         """ """
-        self.parse_features_for_tree(self.new_parent_node(self, ("File Scope",)), file_features)
+        self.parse_features_for_tree(
+            self.new_parent_node(self, ("File Scope",)), file_features
+        )
         if func_features:
-            self.parse_features_for_tree(self.new_parent_node(self, ("Function/Basic Block Scope",)), func_features)
+            self.parse_features_for_tree(
+                self.new_parent_node(self, ("Function/Basic Block Scope",)),
+                func_features,
+            )
         resize_columns_to_content(self.header())
 
     def parse_features_for_tree(self, parent, features):
@@ -1064,24 +1134,32 @@ class CapaExplorerRulegenFeatures(QtWidgets.QTreeWidget):
 
             # level 0
             if type(feature) not in self.parent_items:
-                self.parent_items[type(feature)] = self.new_parent_node(parent, (feature.name.lower(),))
+                self.parent_items[type(feature)] = self.new_parent_node(
+                    parent, (feature.name.lower(),)
+                )
 
             # level 1
             if feature not in self.parent_items:
                 if len(addrs) > 1:
                     self.parent_items[feature] = self.new_parent_node(
-                        self.parent_items[type(feature)], (format_feature(feature),), feature=feature
+                        self.parent_items[type(feature)],
+                        (format_feature(feature),),
+                        feature=feature,
                     )
                 else:
                     self.parent_items[feature] = self.new_leaf_node(
-                        self.parent_items[type(feature)], (format_feature(feature),), feature=feature
+                        self.parent_items[type(feature)],
+                        (format_feature(feature),),
+                        feature=feature,
                     )
 
             # level n > 1
             if len(addrs) > 1:
                 for addr in sorted(addrs):
                     self.new_leaf_node(
-                        self.parent_items[feature], (format_feature(feature), format_address(addr)), feature=feature
+                        self.parent_items[feature],
+                        (format_feature(feature), format_address(addr)),
+                        feature=feature,
                     )
             else:
                 if addrs:
@@ -1136,7 +1214,9 @@ class CapaExplorerQtreeView(QtWidgets.QTreeView):
         self.customContextMenuRequested.connect(self.slot_custom_context_menu_requested)
         self.doubleClicked.connect(self.slot_double_click)
 
-        self.setStyleSheet("QTreeView::item {padding-right: 15 px;padding-bottom: 2 px;}")
+        self.setStyleSheet(
+            "QTreeView::item {padding-right: 15 px;padding-bottom: 2 px;}"
+        )
 
     def reset_ui(self, should_sort=True):
         """reset user interface changes
@@ -1146,7 +1226,10 @@ class CapaExplorerQtreeView(QtWidgets.QTreeView):
         @param should_sort: True, sort results after reset, False don't sort results after reset
         """
         if should_sort:
-            self.sortByColumn(CapaExplorerDataModel.COLUMN_INDEX_RULE_INFORMATION, QtCore.Qt.AscendingOrder)
+            self.sortByColumn(
+                CapaExplorerDataModel.COLUMN_INDEX_RULE_INFORMATION,
+                QtCore.Qt.AscendingOrder,
+            )
 
         self.should_resize_columns = False
         self.expandToDepth(0)
@@ -1338,7 +1421,9 @@ class CapaExplorerQtreeView(QtWidgets.QTreeView):
         column = model_index.column()
         menu = None
 
-        if CapaExplorerDataModel.COLUMN_INDEX_RULE_INFORMATION == column and isinstance(item, CapaExplorerFunctionItem):
+        if CapaExplorerDataModel.COLUMN_INDEX_RULE_INFORMATION == column and isinstance(
+            item, CapaExplorerFunctionItem
+        ):
             # user hovered function item
             menu = self.load_function_item_context_menu(pos, item, model_index)
         else:
@@ -1361,10 +1446,15 @@ class CapaExplorerQtreeView(QtWidgets.QTreeView):
         item = self.map_index_to_source_item(model_index)
         column = model_index.column()
 
-        if CapaExplorerDataModel.COLUMN_INDEX_VIRTUAL_ADDRESS == column and item.location:
+        if (
+            CapaExplorerDataModel.COLUMN_INDEX_VIRTUAL_ADDRESS == column
+            and item.location
+        ):
             # user double-clicked virtual address column - navigate IDA to address
             idc.jumpto(item.location)
 
         if CapaExplorerDataModel.COLUMN_INDEX_RULE_INFORMATION == column:
             # user double-clicked information column - un/expand
-            self.collapse(model_index) if self.isExpanded(model_index) else self.expand(model_index)
+            self.collapse(model_index) if self.isExpanded(model_index) else self.expand(
+                model_index
+            )